# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NiazFinder API",
    "version": "0.2.x-unstable",
    "description": "Generated skeleton of the Next.js App Router API surface. Methods and auth requirements are statically detected; only operations with explicit request/response schemas are stable. Regenerate with `npm run docs:openapi`.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/sasanappstore2/NiazFinder-main/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev (`npm run dev`)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "User token from phone-OTP login (AuthToken)."
      },
      "internalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "INTERNAL_API_SECRET for worker/cron routes. Constant-time compare, fail-closed."
      }
    }
  },
  "paths": {
...
    "/api/users/profile": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "GET /api/users/profile",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "PUT /api/users/profile",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/search": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "GET /api/users/search",
        "description": "Authenticated route (Bearer token). Elevated operations additionally require ADMIN / SUPER_ADMIN role.",
        "responses": {
          "200": {
            "description": "Success (schema not yet documented — see route source)."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Unauthenticated."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}